services/delivery-orchestrator/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
google-cloud-firestore==2.14.0
